+++ src/recipe.rs
@@ -4,17 +4,17 @@ use std::collections::HashMap;
 #[derive(Debug, Deserialize)]
 pub struct GameObject {
     #[serde(rename = "m_FileID")]
-    pub file_id: i32,
+    pub file_id: i64,
     #[serde(rename = "m_PathID")]
-    pub path_id: i32,
+    pub path_id: i64,
 }
 
 #[derive(Debug, Deserialize)]
 pub struct ScriptInfo {
     #[serde(rename = "m_FileID")]
-    pub file_id: i32,
+    pub file_id: i64,
     #[serde(rename = "m_PathID")]
-    pub path_id: i32,
+    pub path_id: i64,
 }
 
 #[derive(Debug, Deserialize)]
@@ -22,27 +22,27 @@ pub struct RecipeItem {
     #[serde(rename = "Name")]
     pub name: String,
     #[serde(rename = "ID")]
-    pub id: i32,
+    pub id: i16,
     #[serde(rename = "SID")]
     pub sid: String,
     #[serde(rename = "Type")]
-    pub type_: i32, // 避讳关键字 "type"
+    pub type_: i64, // 避讳关键字 "type"
     #[serde(rename = "Handcraft")]
     pub handcraft: bool,
     #[serde(rename = "Explicit")]
     pub explicit: bool,
     #[serde(rename = "TimeSpend")]
-    pub time_spend: i32,
+    pub time_spend: i64,
     #[serde(rename = "Items")]
-    pub items: Vec<i32>,
+    pub items: Vec<i64>,
     #[serde(rename = "ItemCounts")]
-    pub item_counts: Vec<i32>,
+    pub item_counts: Vec<i64>,
     #[serde(rename = "Results")]
-    pub results: Vec<i32>,
+    pub results: Vec<i64>,
     #[serde(rename = "ResultCounts")]
-    pub result_counts: Vec<i32>,
+    pub result_counts: Vec<i64>,
     #[serde(rename = "GridIndex")]
-    pub grid_index: i32,
+    pub grid_index: i64,
     #[serde(rename = "IconPath")]
     pub icon_path: String,
     #[serde(rename = "Description")]
@@ -56,7 +56,7 @@ pub struct RecipeProtoSet {
     #[serde(rename = "m_GameObject")]
     pub game_object: GameObject,
     #[serde(rename = "m_Enabled")]
-    pub enabled: i32,
+    pub enabled: i64,
     #[serde(rename = "m_Script")]
     pub script: ScriptInfo,
     #[serde(rename = "m_Name")]
